LPS-62118 extract common logic to a base display context
