+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/display/context/BaseLayoutDisplayContext.java
@@ -1,0 +1,296 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.layout.admin.web.display.context;
+
+import com.liferay.application.list.util.LatentGroupManagerUtil;
+import com.liferay.layout.admin.web.constants.LayoutAdminPortletKeys;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
+import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
+import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.model.Group;
+import com.liferay.portal.model.Layout;
+import com.liferay.portal.model.LayoutConstants;
+import com.liferay.portal.model.LayoutSet;
+import com.liferay.portal.security.permission.ActionKeys;
+import com.liferay.portal.service.LayoutLocalServiceUtil;
+import com.liferay.portal.service.LayoutSetLocalServiceUtil;
+import com.liferay.portal.service.permission.GroupPermissionUtil;
+import com.liferay.portal.theme.ThemeDisplay;
+import com.liferay.portal.util.PortalUtil;
+import com.liferay.portlet.exportimport.staging.StagingUtil;
+
+import javax.portlet.PortletRequest;
+import javax.portlet.PortletURL;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
+/**
+ * @author Julio Camarero
+ */
+public class BaseLayoutDisplayContext {
+
+	public BaseLayoutDisplayContext(
+			LiferayPortletRequest liferayPortletRequest,
+			LiferayPortletResponse liferayPortletResponse)
+		throws PortalException {
+
+		this.liferayPortletRequest = liferayPortletRequest;
+		this.liferayPortletResponse = liferayPortletResponse;
+
+		themeDisplay = (ThemeDisplay)liferayPortletRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+	}
+
+	public PortletURL getAddLayoutURL(long selPlid, Boolean privateLayout) {
+		PortletURL addPagesURL = PortalUtil.getControlPanelPortletURL(
+			liferayPortletRequest, LayoutAdminPortletKeys.GROUP_PAGES,
+			PortletRequest.RENDER_PHASE);
+
+		addPagesURL.setParameter("mvcPath", "/add_layout.jsp");
+
+		if (selPlid >= LayoutConstants.DEFAULT_PLID) {
+			addPagesURL.setParameter("selPlid", String.valueOf(selPlid));
+		}
+
+		addPagesURL.setParameter("groupId", String.valueOf(getSelGroupId()));
+
+		if (privateLayout != null) {
+			addPagesURL.setParameter(
+				"privateLayout", String.valueOf(privateLayout));
+		}
+
+		return addPagesURL;
+	}
+
+	public PortletURL getEditLayoutURL(long selPlid, Boolean privateLayout) {
+		PortletURL editPublicLayoutURL = PortalUtil.getControlPanelPortletURL(
+			liferayPortletRequest, LayoutAdminPortletKeys.GROUP_PAGES,
+			PortletRequest.RENDER_PHASE);
+
+		if (selPlid >= LayoutConstants.DEFAULT_PLID) {
+			editPublicLayoutURL.setParameter(
+				"selPlid", String.valueOf(selPlid));
+		}
+
+		if (privateLayout != null) {
+			editPublicLayoutURL.setParameter(
+				"privateLayout", String.valueOf(privateLayout));
+		}
+
+		Group liveGroup = getLiveGroup();
+
+		editPublicLayoutURL.setParameter(
+			"groupId", String.valueOf(liveGroup.getGroupId()));
+		editPublicLayoutURL.setParameter("viewLayout", Boolean.TRUE.toString());
+
+		return editPublicLayoutURL;
+	}
+
+	public Long getLayoutId() {
+		if (_layoutId != null) {
+			return _layoutId;
+		}
+
+		_layoutId = LayoutConstants.DEFAULT_PARENT_LAYOUT_ID;
+
+		Layout selLayout = getSelLayout();
+
+		if (selLayout != null) {
+			_layoutId = selLayout.getLayoutId();
+		}
+
+		return _layoutId;
+	}
+
+	public String getRootNodeName() {
+		if (_rootNodeName != null) {
+			return _rootNodeName;
+		}
+
+		_rootNodeName = getRootNodeName(isPrivateLayout());
+
+		return _rootNodeName;
+	}
+
+	public String getRootNodeName(boolean privateLayout) {
+		Group liveGroup = getLiveGroup();
+
+		return liveGroup.getLayoutRootNodeName(
+			privateLayout, themeDisplay.getLocale());
+	}
+
+	public long getSelGroupId() {
+		Group selGroup = getSelGroup();
+
+		if (selGroup != null) {
+			return selGroup.getGroupId();
+		}
+
+		return 0;
+	}
+
+	public Layout getSelLayout() {
+		if (_selLayout != null) {
+			return _selLayout;
+		}
+
+		if (getSelPlid() != LayoutConstants.DEFAULT_PLID) {
+			_selLayout = LayoutLocalServiceUtil.fetchLayout(getSelPlid());
+		}
+
+		return _selLayout;
+	}
+
+	public LayoutSet getSelLayoutSet() throws PortalException {
+		if (_selLayoutSet != null) {
+			return _selLayoutSet;
+		}
+
+		Group group = getStagingGroup();
+
+		if (group == null) {
+			group = getLiveGroup();
+		}
+
+		_selLayoutSet = LayoutSetLocalServiceUtil.getLayoutSet(
+			group.getGroupId(), isPrivateLayout());
+
+		return _selLayoutSet;
+	}
+
+	public Long getSelPlid() {
+		if (_selPlid != null) {
+			return _selPlid;
+		}
+
+		_selPlid = ParamUtil.getLong(
+			liferayPortletRequest, "selPlid", LayoutConstants.DEFAULT_PLID);
+
+		return _selPlid;
+	}
+
+	public boolean isPrivateLayout() {
+		if (_privateLayout != null) {
+			return _privateLayout;
+		}
+
+		Group selGroup = getSelGroup();
+
+		if (selGroup.isLayoutSetPrototype() ||
+			selGroup.isLayoutSetPrototype()) {
+
+			_privateLayout = true;
+
+			return _privateLayout;
+		}
+
+		Layout selLayout = getSelLayout();
+
+		if (getSelLayout() != null) {
+			_privateLayout = selLayout.isPrivateLayout();
+
+			return _privateLayout;
+		}
+
+		Layout layout = themeDisplay.getLayout();
+
+		if (!layout.isTypeControlPanel()) {
+			_privateLayout = layout.isPrivateLayout();
+
+			return _privateLayout;
+		}
+
+		_privateLayout = ParamUtil.getBoolean(
+			liferayPortletRequest, "privateLayout");
+
+		return _privateLayout;
+	}
+
+	public boolean isShowAddRootLayoutButton() throws PortalException {
+		return GroupPermissionUtil.contains(
+			themeDisplay.getPermissionChecker(), getSelGroup(),
+			ActionKeys.ADD_LAYOUT);
+	}
+
+	protected HttpServletRequest getHttpServletRequest() {
+		if (_httpServletRequest != null) {
+			return _httpServletRequest;
+		}
+
+		_httpServletRequest = PortalUtil.getHttpServletRequest(
+			liferayPortletRequest);
+
+		return _httpServletRequest;
+	}
+
+	protected HttpSession getHttpSession() {
+		HttpServletRequest httpServletRequest = getHttpServletRequest();
+
+		return httpServletRequest.getSession();
+	}
+
+	protected Group getLiveGroup() {
+		if (_liveGroup != null) {
+			return _liveGroup;
+		}
+
+		_liveGroup = StagingUtil.getLiveGroup(getSelGroupId());
+
+		return _liveGroup;
+	}
+
+	protected Group getSelGroup() {
+		if (_selGroup != null) {
+			return _selGroup;
+		}
+
+		_selGroup = themeDisplay.getScopeGroup();
+
+		if (_selGroup.isControlPanel()) {
+			_selGroup = LatentGroupManagerUtil.getLatentGroup(getHttpSession());
+		}
+
+		return _selGroup;
+	}
+
+	protected Group getStagingGroup() {
+		if (_stagingGroup != null) {
+			return _stagingGroup;
+		}
+
+		_stagingGroup = StagingUtil.getStagingGroup(getSelGroupId());
+
+		return _stagingGroup;
+	}
+
+	protected final LiferayPortletRequest liferayPortletRequest;
+	protected final LiferayPortletResponse liferayPortletResponse;
+	protected final ThemeDisplay themeDisplay;
+
+	private HttpServletRequest _httpServletRequest;
+	private Long _layoutId;
+	private Group _liveGroup;
+	private Boolean _privateLayout;
+	private String _rootNodeName;
+	private Group _selGroup;
+	private Layout _selLayout;
+	private LayoutSet _selLayoutSet;
+	private Long _selPlid;
+	private Group _stagingGroup;
+
+}
